back: API 추가
환자가 (자신이) 테스트한 재활코스 전체를 확인하는 API
의사가 본인이 등록한 영상 삭제하는 API

+++ back/api/FileViews.py
@@ -112,7 +112,11 @@ class FileUpload(APIView):
 class FileDelete(APIView):
     def post(self, request):
         body = json.loads(request.body.decode('utf-8'))
-        data = Correctpic.objects.filter(exercisename=body['name']).first()
+        data = Correctpic.objects.filter(exercisetype=body['type'], exercisename=body['name']).first()
+
+        if Patientpic.objects.filter(correctpicid=data.uid).first() is not None:
+            raise AuthenticationFailed("파일을 삭제할 수 없습니다.")
+
         data.delete()
 
         response = Response()
